app/requirements.txt
fastapi
uvicorn
pyjwt
google-generativeai
uvicorn